review: move tests into bun-create.test.ts, assert --bun hoist, document -- separator
- move the argv-forwarding tests into test/cli/install/bun-create.test.ts
  alongside the other `bun create` coverage
- build the fixture tarball via `pack()` from harness (bun pm pack) instead
  of a hand-rolled tar header writer
- have the fixture report `!!process.versions.bun` and assert it in the
  `--bun` cases, so the hoist onto bunx is covered independently of the
  argv filter
- document the `--` separator in docs/runtime/templating/create.mdx
